+++ src/index.tsx
@@ -5,7 +5,7 @@ import { createRoot } from 'react-dom/client';
 import { Routes } from './routes/routes';
 import { store } from './store/index';
 
-import './assets/style/index.scss';
+import './style/index.scss';
 
 const container = document.getElementById('root');
 if (!container) throw new Error('Failed to find the root element');
