+++ src/style/_variable/_color.scss
@@ -5,7 +5,10 @@ $c_light: #fff;
 
 $c_current: #36c;
 $c_accent: #f06;
+$c_sub: #36c6;
 $c_bg: #f2f2f2;
+$c_disabled: #ccc;
 
 $c_shadow: #0009;
 $c_screen: #0009;
+
