+++ src/style/_variable/_font.scss
@@ -1,0 +1,2 @@
+$f_regular: 400;
+$f_bold: 700;
