+++ src/style/_variable/_index.scss
@@ -4,4 +4,5 @@
 @forward 'mixin/rect';
 @forward 'mixin/tracking';
 @forward 'color';
+@forward 'font';
 @forward 'transition';
