+++ src/style/layout/dialog.module.scss
